Track service managements(deploy/modify/destroy/rollback/retry/purge) into table service_order

+++ modules/deployment/src/main/java/org/eclipse/xpanse/modules/deployment/deployers/terraform/terraformlocal/TerraformLocalDeployment.java
@@ -19,7 +19,6 @@ import java.util.Comparator;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Set;
 import java.util.UUID;
 import java.util.concurrent.Executor;
 import java.util.stream.Stream;
@@ -27,8 +26,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.eclipse.xpanse.modules.database.service.DeployServiceEntity;
 import org.eclipse.xpanse.modules.deployment.DeployServiceEntityHandler;
+import org.eclipse.xpanse.modules.deployment.deployers.opentofu.exceptions.OpenTofuExecutorException;
 import org.eclipse.xpanse.modules.deployment.deployers.terraform.callbacks.TerraformDeploymentResultCallbackManager;
-import org.eclipse.xpanse.modules.deployment.deployers.terraform.exceptions.TerraformExecutorException;
 import org.eclipse.xpanse.modules.deployment.deployers.terraform.terraformboot.TerraformBootDeployment;
 import org.eclipse.xpanse.modules.deployment.deployers.terraform.terraformboot.generated.model.TerraformResult;
 import org.eclipse.xpanse.modules.deployment.deployers.terraform.terraformlocal.config.TerraformLocalConfig;
@@ -43,6 +42,7 @@ import org.eclipse.xpanse.modules.orchestrator.deployment.DeployResult;
 import org.eclipse.xpanse.modules.orchestrator.deployment.DeployTask;
 import org.eclipse.xpanse.modules.orchestrator.deployment.Deployer;
 import org.eclipse.xpanse.modules.orchestrator.deployment.DeploymentScriptValidationResult;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.stereotype.Component;
 
@@ -56,7 +56,8 @@ public class TerraformLocalDeployment implements Deployer {
     public static final String SCRIPT_FILE_NAME = "resources.tf";
     public static final String STATE_FILE_NAME = "terraform.tfstate";
     public static final String TF_DEBUG_FLAG = "TF_LOG";
-
+    @Value("${clean.workspace.after.deployment.enabled:true}")
+    private boolean cleanWorkspaceAfterDeploymentEnabled;
     @Resource
     private TerraformInstaller terraformInstaller;
     @Resource
@@ -83,7 +84,6 @@ public class TerraformLocalDeployment implements Deployer {
     public DeployResult deploy(DeployTask task) {
         DeployResult deployResult = new DeployResult();
         deployResult.setOrderId(task.getOrderId());
-        deployResult.setServiceId(task.getServiceId());
         asyncExecDeploy(task);
         return deployResult;
     }
@@ -107,7 +107,6 @@ public class TerraformLocalDeployment implements Deployer {
         }
 
         DeployResult destroyResult = new DeployResult();
-        destroyResult.setServiceId(task.getServiceId());
         destroyResult.setOrderId(task.getOrderId());
         asyncExecDestroy(task, resourceState);
         return destroyResult;
@@ -130,17 +129,14 @@ public class TerraformLocalDeployment implements Deployer {
             throw new ServiceNotDeployedException(errorMsg);
         }
         DeployResult modifyResult = new DeployResult();
-        modifyResult.setServiceId(task.getServiceId());
         modifyResult.setOrderId(task.getOrderId());
         asyncExecModify(task, resourceState);
         return modifyResult;
     }
 
     private void asyncExecDeploy(DeployTask task) {
-        String workspace = getWorkspacePath(task.getServiceId());
-        // Create the workspace.
-        buildWorkspace(workspace);
-        prepareDeployWorkspaceWithScripts(task, workspace);
+        String workspace = createWorkspaceForTask(task.getOrderId());
+        prepareDeploymentScripts(workspace, task.getOcl().getDeployment(), null);
         TerraformLocalExecutor executor = getExecutorForDeployTask(task, workspace, true);
         // Execute the terraform command asynchronously.
         taskExecutor.execute(() -> {
@@ -159,15 +155,17 @@ public class TerraformLocalDeployment implements Deployer {
                     terraformInstaller.getExactVersionOfTerraform(executor.getExecutorPath()));
             Map<String, String> importantFileContentMap = executor.getImportantFilesContent();
             terraformResult.setImportantFileContentMap(importantFileContentMap);
-            terraformDeploymentResultCallbackManager.deployCallback(task.getServiceId(),
+            terraformDeploymentResultCallbackManager.orderCallback(task.getOrderId(),
                     terraformResult);
-            deleteStoredFiles(workspace, importantFileContentMap.keySet());
+            if (cleanWorkspaceAfterDeploymentEnabled) {
+                deleteWorkspace(workspace);
+            }
         });
     }
 
     private void asyncExecDestroy(DeployTask task, String tfState) {
-        String workspace = getWorkspacePath(task.getServiceId());
-        prepareDestroyWorkspaceWithScripts(task, workspace, tfState);
+        String workspace = createWorkspaceForTask(task.getOrderId());
+        prepareDeploymentScripts(workspace, task.getOcl().getDeployment(), tfState);
         TerraformLocalExecutor executor = getExecutorForDeployTask(task, workspace, false);
         // Execute the terraform command asynchronously.
         taskExecutor.execute(() -> {
@@ -186,16 +184,17 @@ public class TerraformLocalDeployment implements Deployer {
                     terraformInstaller.getExactVersionOfTerraform(executor.getExecutorPath()));
             Map<String, String> importantFileContentMap = executor.getImportantFilesContent();
             terraformResult.setImportantFileContentMap(importantFileContentMap);
-            terraformDeploymentResultCallbackManager.destroyCallback(task.getServiceId(),
-                    terraformResult, task.getDeploymentScenario());
-            deleteStoredFiles(workspace, importantFileContentMap.keySet());
+            terraformDeploymentResultCallbackManager.orderCallback(task.getOrderId(),
+                    terraformResult);
+            if (cleanWorkspaceAfterDeploymentEnabled) {
+                deleteWorkspace(workspace);
+            }
         });
     }
 
     private void asyncExecModify(DeployTask task, String tfState) {
-        String workspace = getWorkspacePath(task.getServiceId());
-        prepareDestroyWorkspaceWithScripts(task, workspace, tfState);
-        prepareDeployWorkspaceWithScripts(task, workspace);
+        String workspace = createWorkspaceForTask(task.getOrderId());
+        prepareDeploymentScripts(workspace, task.getOcl().getDeployment(), tfState);
         TerraformLocalExecutor executor = getExecutorForDeployTask(task, workspace, true);
         // Execute the terraform command asynchronously.
         taskExecutor.execute(() -> {
@@ -214,27 +213,21 @@ public class TerraformLocalDeployment implements Deployer {
                     terraformInstaller.getExactVersionOfTerraform(executor.getExecutorPath()));
             Map<String, String> importantFileContentMap = executor.getImportantFilesContent();
             terraformResult.setImportantFileContentMap(importantFileContentMap);
-            terraformDeploymentResultCallbackManager.modifyCallback(task.getServiceId(),
+            terraformDeploymentResultCallbackManager.orderCallback(task.getOrderId(),
                     terraformResult);
-            deleteStoredFiles(workspace, importantFileContentMap.keySet());
+            if (cleanWorkspaceAfterDeploymentEnabled) {
+                deleteWorkspace(workspace);
+            }
         });
     }
 
     @Override
     public String getDeploymentPlanAsJson(DeployTask task) {
-        String workspace = getWorkspacePath(task.getServiceId());
-        // Create the workspace.
-        buildWorkspace(workspace);
-        prepareDeployWorkspaceWithScripts(task, workspace);
+        String workspace = createWorkspaceForTask(task.getOrderId());
+        prepareDeploymentScripts(workspace, task.getOcl().getDeployment(), null);
         // Execute the terraform command.
         TerraformLocalExecutor executor = getExecutorForDeployTask(task, workspace, true);
         return executor.getTerraformPlanAsJson();
-    }
-
-    @Override
-    public void deleteTaskWorkspace(UUID taskId) {
-        String workspace = getWorkspacePath(taskId);
-        deleteWorkspace(workspace);
     }
 
     /**
@@ -247,18 +240,6 @@ public class TerraformLocalDeployment implements Deployer {
         } catch (Exception e) {
             log.error("Delete workspace:{} error.", workspace, e);
         }
-    }
-
-    private void deleteStoredFiles(String workspace, Set<String> fileNames) {
-        fileNames.forEach(fileName -> {
-            try {
-                String path = workspace + File.separator + fileName;
-                File file = new File(path);
-                Files.deleteIfExists(file.toPath());
-            } catch (IOException e) {
-                log.error("Delete file with name:{} error.", fileName, e);
-            }
-        });
     }
 
     /**
@@ -290,99 +271,52 @@ public class TerraformLocalDeployment implements Deployer {
                 getSubDirectory(deployment), deployResultFileUtils);
     }
 
-    private void prepareDeployWorkspaceWithScripts(DeployTask deployTask, String workspace) {
-        if (Objects.nonNull(deployTask.getOcl().getDeployment().getDeployer())) {
-            createScriptFile(workspace, deployTask.getOcl().getDeployment().getDeployer());
+    private String createWorkspaceForTask(UUID orderId) {
+        String workspace = terraformLocalConfig.getWorkspaceDirectory() + File.separator + orderId;
+        File ws = new File(System.getProperty("java.io.tmpdir"), workspace);
+        if (!ws.exists() && !ws.mkdirs()) {
+            throw new OpenTofuExecutorException(
+                    "Create workspace failed, File path not created: " + ws.getAbsolutePath());
         }
-        if (Objects.nonNull(deployTask.getOcl().getDeployment().getScriptsRepo())) {
-            scriptsGitRepoManage.checkoutScripts(workspace,
-                    deployTask.getOcl().getDeployment().getScriptsRepo());
-        }
+        return ws.getAbsolutePath();
     }
 
-    private void prepareDestroyWorkspaceWithScripts(DeployTask deployTask, String workspace,
-                                                    String tfState) {
-        log.info("start create terraform destroy workspace and script");
-        File parentPath = new File(workspace);
-        if (!parentPath.exists() || !parentPath.isDirectory()) {
-            parentPath.mkdirs();
-        }
-        if (Objects.nonNull(deployTask.getOcl().getDeployment().getDeployer())) {
-            createDestroyScriptFile(workspace, tfState);
-        } else if (Objects.nonNull(deployTask.getOcl().getDeployment().getScriptsRepo())) {
-            scriptsGitRepoManage.checkoutScripts(workspace,
-                    deployTask.getOcl().getDeployment().getScriptsRepo());
-            String scriptPath = workspace + File.separator + deployTask.getOcl().getDeployment()
-                    .getScriptsRepo().getScriptsPath() + File.separator + STATE_FILE_NAME;
-            try (FileWriter scriptWriter = new FileWriter(scriptPath)) {
-                scriptWriter.write(tfState);
-            } catch (IOException e) {
-                throw new RuntimeException(e);
+    private void prepareDeploymentScripts(String workspace, Deployment deployment, String tfState) {
+        if (Objects.nonNull(deployment.getDeployer())) {
+            createScriptFile(workspace, deployment.getDeployer());
+            if (StringUtils.isNotBlank(tfState)) {
+                createServiceStateFile(workspace, tfState);
+            }
+        } else if (Objects.nonNull(deployment.getScriptsRepo())) {
+            scriptsGitRepoManage.checkoutScripts(workspace, deployment.getScriptsRepo());
+            String scriptPath =
+                    workspace + File.separator + deployment.getScriptsRepo().getScriptsPath();
+            if (StringUtils.isNotBlank(tfState)) {
+                createServiceStateFile(scriptPath, tfState);
             }
         }
     }
 
-    /**
-     * Create terraform script.
-     *
-     * @param workspace the workspace for terraform.
-     * @param script    terraform scripts of the task.
-     */
-    private void createScriptFile(String workspace, String script) {
-        log.info("start create terraform script");
-        String scriptPath = workspace + File.separator + SCRIPT_FILE_NAME;
+    private void createScriptFile(String path, String scriptContext) {
+        String scriptPath = path + File.separator + SCRIPT_FILE_NAME;
         try (FileWriter scriptWriter = new FileWriter(scriptPath)) {
-            scriptWriter.write(script);
-            log.info("Terraform script create success");
+            scriptWriter.write(scriptContext);
+            log.info("Create script file success.");
         } catch (IOException ex) {
-            log.error("create version file failed.", ex);
-            throw new TerraformExecutorException("create version file failed.", ex);
+            log.error("Create script file success.", ex);
+            throw new OpenTofuExecutorException("Create script file success.", ex);
         }
     }
 
-    /**
-     * Create terraform workspace and script.
-     *
-     * @param workspace the workspace for terraform.
-     * @param tfState   terraform file tfstate of the task.
-     */
-    private void createDestroyScriptFile(String workspace, String tfState) {
-
-        String scriptPath = workspace + File.separator + STATE_FILE_NAME;
-        try (FileWriter scriptWriter = new FileWriter(scriptPath)) {
-            scriptWriter.write(tfState);
-            log.info("create terraform destroy workspace and script success.");
+    private void createServiceStateFile(String path, String tfStateContext) {
+        String tfStateFileName = path + File.separator + STATE_FILE_NAME;
+        try (FileWriter scriptWriter = new FileWriter(tfStateFileName)) {
+            scriptWriter.write(tfStateContext);
+            log.info("Create service state file success.");
         } catch (IOException e) {
-            log.error("create terraform destroy workspace and script failed.", e);
-            throw new TerraformExecutorException(
-                    "create terraform destroy workspace and script failed.", e);
+            log.error("Create service state file failed.", e);
+            throw new OpenTofuExecutorException("Create service state file failed.", e);
         }
-
-    }
-
-    /**
-     * Build workspace of the `terraform`.
-     *
-     * @param workspace The workspace of the task.
-     */
-    private void buildWorkspace(String workspace) {
-        log.info("start creating workspace");
-        File ws = new File(workspace);
-        if (!ws.exists() && !ws.mkdirs()) {
-            throw new TerraformExecutorException(
-                    "Create workspace failed, File path not created: " + ws.getAbsolutePath());
-        }
-        log.info("workspace create success,Working directory is " + ws.getAbsolutePath());
-    }
-
-    /**
-     * Get the workspace path for terraform.
-     *
-     * @param taskId The id of the task.
-     */
-    private String getWorkspacePath(UUID taskId) {
-        return System.getProperty("java.io.tmpdir") + File.separator
-                + terraformLocalConfig.getWorkspaceDirectory() + File.separator + taskId.toString();
     }
 
 
@@ -399,14 +333,8 @@ public class TerraformLocalDeployment implements Deployer {
      */
     @Override
     public DeploymentScriptValidationResult validate(Deployment deployment) {
-        String workspace = getWorkspacePath(UUID.randomUUID());
-        // Create the workspace.
-        buildWorkspace(workspace);
-        if (Objects.nonNull(deployment.getDeployer())) {
-            createScriptFile(workspace, deployment.getDeployer());
-        } else {
-            scriptsGitRepoManage.checkoutScripts(workspace, deployment.getScriptsRepo());
-        }
+        String workspace = createWorkspaceForTask(UUID.randomUUID());
+        prepareDeploymentScripts(workspace, deployment, null);
         TerraformLocalExecutor executor =
                 getExecutor(new HashMap<>(), new HashMap<>(), workspace, deployment);
         DeploymentScriptValidationResult validationResult = executor.tfValidate();
